feat: reorganização do sidebar e integração do public/users/:id
