+++ package.json
@@ -16,6 +16,7 @@
     "@emotion/styled": "^11.13.0",
     "axios": "^1.7.7",
     "framer-motion": "^11.5.4",
+    "jwt-decode": "^4.0.0",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
     "react-hook-form": "^7.53.1",
