+++ src/service/Auth.ts
@@ -1,6 +1,5 @@
-import { UserForgot, UserSignIn, UserSingUp } from "../interface/UserInterface";
+import { UserForgot, UserSignIn, UserSingUp, UserResponse } from "../interface/UserInterface";
 import { apiAuth } from "./api";
-
 
 const signUp = async (data: UserSingUp) => {
     const response = await apiAuth.post(`/public/users/register`, data);
@@ -8,7 +7,7 @@ const signUp = async (data: UserSingUp) => {
 };
 
 const signIn = async (data: UserSignIn) => {
-    const response = await apiAuth.post(`public/auth`, data);
+    const response = await apiAuth.post(`/public/auth`, data);
     return response;
 };
 
@@ -17,8 +16,21 @@ const forgot = async (data: UserForgot) => {
     return response;
 };
 
-export{
+const fetchUserData = async (): Promise<UserResponse> => {
+    const id = localStorage.getItem('id');
+
+    if (!id) {
+        throw new Error('ID não foi encontrado'); 
+    }
+    const response = await apiAuth.get<UserResponse>(`/public/users/${id}`);
+    
+    return response.data; 
+};
+
+
+export {
     signUp,
     signIn,
-    forgot
+    forgot,
+    fetchUserData
 };
